requirements.txt
psutil
Pillow
sv-ttk
pywinstyles